0requirements.txt
segmentation-models==1.0.1
efficientnet==1.0.0
gdown
image-classifiers==1.0.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib
matplotlib-inline
numpy
Pillow
pycocotools
pycotools
requests
requests-oauthlib
scikit-image
scikit-learn
seaborn
tensorflow==2.8.0
tensorflow-intel
tensorflow-io-gcs-filesystem
protobuf==3.20.*